refactor: incapsulate all localization setup logic in one place
